requirements.txt
# Core LLM & Agent Framework
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Vector Store
chromadb>=0.4.0

# Web Search
tavily-python>=0.3.0

# Mathematical Computation
wolframalpha>=5.0.0

# PDF & Image Processing
pymupdf>=1.24.0
pillow>=10.0.0

# Web UI
streamlit>=1.35.0

# Utilities
python-dotenv>=1.0.0
tiktoken>=0.7.0

# Evaluation & Analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

